backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
exifread==3.0.0
piexif==1.1.3
ffmpeg-python==0.2.0
reportlab==4.0.7
pydantic==2.5.0
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
